Extract vim interface & split ActionJob.
